--- x-pack/plugins/triggers_actions_ui/public/application/lib/action_connector_api/action_connector_api.test.ts
@@ -1,190 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { ActionConnectorWithoutId, ActionType } from '../../../types';
-import { httpServiceMock } from '../../../../../../../src/core/public/mocks';
-import {
-  createActionConnector,
-  deleteActions,
-  loadActionTypes,
-  loadAllActions,
-  updateActionConnector,
-  executeAction,
-} from './index';
-
-const http = httpServiceMock.createStartContract();
-
-beforeEach(() => jest.resetAllMocks());
-
-describe('loadActionTypes', () => {
-  test('should call get types API', async () => {
-    const apiResponseValue = [
-      {
-        id: 'test',
-        name: 'Test',
-        enabled: true,
-        enabled_in_config: true,
-        enabled_in_license: true,
-        minimum_license_required: 'basic',
-      },
-    ];
-    http.get.mockResolvedValueOnce(apiResponseValue);
-
-    const resolvedValue: ActionType[] = [
-      {
-        id: 'test',
-        name: 'Test',
-        enabled: true,
-        enabledInConfig: true,
-        enabledInLicense: true,
-        minimumLicenseRequired: 'basic',
-      },
-    ];
-
-    const result = await loadActionTypes({ http });
-    expect(result).toEqual(resolvedValue);
-    expect(http.get.mock.calls[0]).toMatchInlineSnapshot(`
-      Array [
-        "/api/actions/connector_types",
-      ]
-    `);
-  });
-});
-
-describe('loadAllActions', () => {
-  test('should call getAll actions API', async () => {
-    http.get.mockResolvedValueOnce([]);
-
-    const result = await loadAllActions({ http });
-    expect(result).toEqual([]);
-    expect(http.get.mock.calls[0]).toMatchInlineSnapshot(`
-      Array [
-        "/api/actions/connectors",
-      ]
-    `);
-  });
-});
-
-describe('createActionConnector', () => {
-  test('should call create action API', async () => {
-    const apiResponse = {
-      connector_type_id: 'test',
-      is_preconfigured: false,
-      name: 'My test',
-      config: {},
-      secrets: {},
-      id: '123',
-    };
-    http.post.mockResolvedValueOnce(apiResponse);
-
-    const connector: ActionConnectorWithoutId<{}, {}> = {
-      actionTypeId: 'test',
-      isPreconfigured: false,
-      name: 'My test',
-      config: {},
-      secrets: {},
-    };
-    const resolvedValue = { ...connector, id: '123' };
-
-    const result = await createActionConnector({ http, connector });
-    expect(result).toEqual(resolvedValue);
-    expect(http.post.mock.calls[0]).toMatchInlineSnapshot(`
-      Array [
-        "/api/actions/connector",
-        Object {
-          "body": "{\\"name\\":\\"My test\\",\\"config\\":{},\\"secrets\\":{},\\"connector_type_id\\":\\"test\\",\\"is_preconfigured\\":false}",
-        },
-      ]
-    `);
-  });
-});
-
-describe('updateActionConnector', () => {
-  test('should call the update API', async () => {
-    const id = '123';
-    const apiResponse = {
-      connector_type_id: 'test',
-      is_preconfigured: false,
-      name: 'My test',
-      config: {},
-      secrets: {},
-      id,
-    };
-    http.put.mockResolvedValueOnce(apiResponse);
-
-    const connector: ActionConnectorWithoutId<{}, {}> = {
-      actionTypeId: 'test',
-      isPreconfigured: false,
-      name: 'My test',
-      config: {},
-      secrets: {},
-    };
-    const resolvedValue = { ...connector, id };
-
-    const result = await updateActionConnector({ http, connector, id });
-    expect(result).toEqual(resolvedValue);
-    expect(http.put.mock.calls[0]).toMatchInlineSnapshot(`
-      Array [
-        "/api/actions/connector/123",
-        Object {
-          "body": "{\\"name\\":\\"My test\\",\\"config\\":{},\\"secrets\\":{}}",
-        },
-      ]
-    `);
-  });
-});
-
-describe('deleteActions', () => {
-  test('should call delete API per action', async () => {
-    const ids = ['1', '2', '3'];
-
-    const result = await deleteActions({ ids, http });
-    expect(result).toEqual({ errors: [], successes: [undefined, undefined, undefined] });
-    expect(http.delete.mock.calls).toMatchInlineSnapshot(`
-      Array [
-        Array [
-          "/api/actions/connector/1",
-        ],
-        Array [
-          "/api/actions/connector/2",
-        ],
-        Array [
-          "/api/actions/connector/3",
-        ],
-      ]
-    `);
-  });
-});
-
-describe('executeAction', () => {
-  test('should call execute API', async () => {
-    const id = '123';
-    const params = {
-      stringParams: 'someString',
-      numericParams: 123,
-    };
-
-    http.post.mockResolvedValueOnce({
-      connector_id: id,
-      status: 'ok',
-    });
-
-    const result = await executeAction({ id, http, params });
-    expect(result).toEqual({
-      actionId: id,
-      status: 'ok',
-    });
-    expect(http.post.mock.calls[0]).toMatchInlineSnapshot(`
-      Array [
-        "/api/actions/connector/123/_execute",
-        Object {
-          "body": "{\\"params\\":{\\"stringParams\\":\\"someString\\",\\"numericParams\\":123}}",
-        },
-      ]
-    `);
-  });
-});
