+++ x-pack/plugins/triggers_actions_ui/public/application/lib/action_connector_api/delete.test.ts
@@ -1,0 +1,35 @@
+/*
+ * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+ * or more contributor license agreements. Licensed under the Elastic License
+ * 2.0; you may not use this file except in compliance with the Elastic License
+ * 2.0.
+ */
+
+import { httpServiceMock } from '../../../../../../../src/core/public/mocks';
+import { deleteActions } from './index';
+
+const http = httpServiceMock.createStartContract();
+
+beforeEach(() => jest.resetAllMocks());
+
+describe('deleteActions', () => {
+  test('should call delete API per action', async () => {
+    const ids = ['1', '2', '3'];
+
+    const result = await deleteActions({ ids, http });
+    expect(result).toEqual({ errors: [], successes: [undefined, undefined, undefined] });
+    expect(http.delete.mock.calls).toMatchInlineSnapshot(`
+      Array [
+        Array [
+          "/api/actions/connector/1",
+        ],
+        Array [
+          "/api/actions/connector/2",
+        ],
+        Array [
+          "/api/actions/connector/3",
+        ],
+      ]
+    `);
+  });
+});
